Refactor script loading (fixes remembering of Google Maps as preferred)
Load all scripts in inject.js in correct dependency order so that we can
be certain that L.Google is loaded and we can restore the preferred map.

A nice benefit is that we no longer need the baseURL hack in inject.js
as inject.js is now only used for the browser extension and is
functionally equivalent to what the bookmarklet and greasemonkey
entrypoints do.

Less fortunate is that all this renaming breaks existing
greasemonkey/bookmarklet installations. :-(