layers: Fix mapy.cz Aerial
The /bing/ URIs now fail with 403 unless requested with

    Referer: https://mapy.cz/

This is likely a licensing issue… :-/

We could probably work around this using
https://developer.mozilla.org/en-US/Add-ons/WebExtensions/API/webRequest/onBeforeSendHeaders,
but it'd be wrong and I'm not sure if worth the trouble. More info here
just in case:

* https://stackoverflow.com/a/50013180/3407728
* https://developer.chrome.com/docs/extensions/reference/webRequest/

Alternatively, we can do something like Leaflet.GoogleMutant.js with
http://api.mapy.cz/, then it'd likely be legal, but definitely not worth
the trouble unless someone's willing to fund that development.